refactor: PostDetailNavMore 구현으로 navMore 로직 분리
